invoke sphinx-build-3 unconditionally
Previously I had set SPHINXBUILD=sphinx-build-3 for the spec file but
left the default as sphinx-build, but that means the docs break in
Jenkins.

We *probably* want to unconditionally always build the docs using
Python 3 now... but I am not totally sure this is the right solution. At
some point in the (very near?) future, sphinx-build should become the
Python 3 version anyway.