Drop KERBEROS_HTTP_HOST configuration handle
In 7288d464a2de48c49a7ca290aa4a5fa453baa4e1, it was assumed that a
default realm was configured on clients when it was not.  This caused
lookup failures because GSSAPI canonicalized to an empty realm string.
However, since GSSAPI will automatically select the client-requested
principal if it exists in the ccache, there's no need for the name
logic at all.