new POST /api/v1.0/waivers/+filtered endpoint
The POST /api/v1.0/waivers/+by-subjects-and-testcases endpoint has
proven confusing and difficult to use in practice.

This deprecates that endpoint in favour of a general "filtering" POST
endpoint. It should be easier to grok since the filter parameters work
the same way as they do for GET /api/v1.0/waivers/.