Change ENTRYPOINT in Dockerfile to CMD
This is not a strong opinion but I would recommend to move the command
of starting WaiverDB service in the `Dockerfile` from `ENTRYPOINT` to `CMD`
for pursuing the best practices, so that `ENTRYPOINT` can be reserved for container initialization,
like confiurating from environment variables, waiting for services,
generating keys, or reaping zombie processes.

For OpenShift users who overwrite this command, they should overwrite
the `args` parameter instead of `command` parameter after this change.