Enable loading configs from file when testing
In conftest.py, when the application to be used for testing
was created, create_app() was called with the config_object
waiverdb.config.TestConfig.

The above allowed no custumization of the test configration
by using a configuration file.

At the same time, load_config was providing a way to tell, that
the current environment is a test environment.

In order to be able to use other databases than a local one when
running the unit tests, this change sets the 'TEST' env var
to 'true' and calls create_app() without a parameter, so that
configuration files are considered.

Setting the silent argument to True when calling from_pyfile()
allows ignoring missing cofig files when the environment is DEV or
TEST.

Signed-off-by: Hunor Csomortáni <csomh@redhat.com>