Add parameter for custom registry/repo to the test template
This template will be used in the pipeline so that an intermediate image
can be used for test environment setup.
The default value is still pointing to the official registry.

Also add CPU and memory limits for the created pods, otherwise we are
easily to hit the current UpShift quota.