CI/CD: add C3I jobs to poll for master branch and PRs on Pagure
Two Jenkins jobs are added to poll for master branch changes and new PRs
on Pagure.

As a temporary solution for triggering pipeline builds automatically
when changes on master branch or pull-requests, polling is used before
we eventually move to UMB-WebHook triggers.

The 2 polling jobs are needed because

1. to keep the logic of polling separated with dev-pipeline job, which
could be triggered by UMB-WebHook eventually.
2. Currently Jenkins requires a persist workspace to poll for wildcard branch names.
dev-pipeline job runs Jenkins slaves as pods, which are discarded after
job runs.
3. We need to run dev-pipline for both pre-merge and post-merge
workflows. Polling for PRs and the master branch in a single Jenkins job
will *NOT* trigger a master branch build when a fast-forward merge happens to a PR.

We can't move the Jenkinsfile to an external file otherwise Jenkins will
also poll for the branch where Jenkins file is referenced.