Revert "CI/CD: Install build dependencies at runtime"
This reverts commit 225cd5ec2edc7683d377712b2e577eaa4a006e9c.

I did try running sudo on OpenShift and found it worked (maybe I logged into another cluster with looser policies applied?). But on UpShift Prod it is not allowed to run setuid programs (including sudo). So please revert this change.

Dependencies rarely change. We could introduce some manual operations to deal with that before there is a real solution.
So let's just rebuild the Jenkins slave image with new dependencies.