CI/CD: use commit ID to track a commit in post-merge workflow
Due to #244, we switched to branch name to tracking a commit during a
pipeline build. This leads to a risk for a post-merge build: Because
multiple checkouts are done during a pipeline build, it is possible to
get different commits if master branch advances. So we should still use
commit ID to track a commit in the post-merge workflow.