Improve authentication error response
When using Kerberos authentication it is necessary to configure
"dns_canonicalize_hostname" inside the Kerberos configuration file
otherwise the user will get an error response that is a bit
ambiguous. There's no way to detect if that was the problem or it
is just a generic authentication error. But we can provide an hint
to the user, and also advise to check the doc.