Trigger integration tests on repotracker messages
This PR introduces a new job that starts an integration test automatically
when the image with :latest or :stage is changed. After the test is complete, the result is sent to ResultsDB for Greenwave to decide if it is good for promotion.

The job is configured with CIBuildTrigger supported by the [JMS Messaging plugin][2], which consumes the CI message sent by [repotracker][1].

[1]: https://github.com/release-engineering/repotracker
[2]: https://github.com/jenkinsci/jms-messaging-plugin