CI/CD: Install deps from ./requirements.txt before running unit tests
Generally new dependencies added to waiverdb.spec are not installed
to C3I Jenkins slave pod because of the limitation of OpenShift that no root access to pod.
To apply the new set of dependencies, we have to rebuild the Jenkins slave pod image after
the new waiverdb.spec is merged to master.

This PR tries to mitigate this issue by installing additional PyPI dependencies to user directory
`pip3 install --user -r ./requirements.txt`. This shouldn't have side effect because it doesn't update
RPM installed packages unless `-U` is specified.