CI/CD: flag commit for post-merge build
Follow up https://pagure.io/waiverdb/issue/303:

When a post-merge build starts, the commit is marked as 'pending'.
When a post-merge build succeeds, the commit is marked as 'success'.
When a post-merge build fails, the commit is marked as 'failure'.

This PR depends on https://pagure.io/c3i-library/pull-request/10.