CI/CD: inject full integration test to promotion workflows
This change will run the full integration test in the dev->stage and stage->prod promotion workflows. The old functional test will still be used for gating the dev image.

In the short term, we are running the Factory 2.0 integration test in the PSI `c3i` project rather than the `waiverdb-test` project.