waivers/+filtered API endpoint: handle scenarios (#431)
The FilteredWaiversResource method which backs this endpoint
tries to find only 'current' waivers by grouping all waivers
by subject_type, subject_identifier and testcase, then taking
only the highest-numbered waiver from each group. This breaks if
we have multiple waivers for the same subject and testcase but
with different scenarios - the endpoint will only return one
waiver, the one with the highest ID, and will not return the
valid and current waivers for other scenarios as it should.

Signed-off-by: Adam Williamson <awilliam@redhat.com>